Move the bulk of the main logic into its own function
This takes the for-loop that clones and fetches repositories and
stuffs it into its own function.  The rest of the main code is broken
off with a larger comment block, and then that new function is used to
reduce the amount of ‘main’ code.

However, there is one slight behavioral change in this patch.  The
output would previously say

    Cloning or fetching $user\'s…

The $user has been removed from that message because the variable is
not defined before that function.  I felt it was easier to just remove
the user name than to make $user global or pass it as an extra
argument, under the assumption that the username will be directly
above anyways from where the program was just run.