Fix the ‘cannot open <user> bug’
Commit 11096bea (‘Implement the --list command’) introduced a second,
optional command-line parameter.  What was once

    my $user = shift @ARGV;

turned into

    my ($user, $command) = @ARGV;

This broke the call prompt() inside of clone_or_fetch() because
IO::Prompt will attempt to read input from files listed in @ARGV if
they are present.  Before we shifted the value for $user directly out
of @ARGV, which left the array empty and thus IO::Prompt would read
from standard input.  However, commit 11096bea removed the call to
shift.  This would leave the username as the first element of @ARGV,
and IO::Prompt would die trying to read from a file with that name.

The problem can be fixed by shifting everything out of @ARGV to ensure
that it is empty by the time prompt() is called.